Rename dirs/files according to the hashicorp providers' guidelines
