+++ main.go
@@ -4,7 +4,7 @@ import (
 	"github.com/hashicorp/terraform/plugin"
 	"github.com/hashicorp/terraform/terraform"
 	"github.com/vmware/terraform-provider-vra7/utils"
-	"github.com/vmware/terraform-provider-vra7/vrealize"
+	"github.com/vmware/terraform-provider-vra7/vra7"
 )
 
 func main() {
