+++ sdk/vra7_sdk_test.go
